test(cypress): extract tests for configure-angular-json and write-cypress-files rules
